fix(datepickers): stop calling onChange from within DatePicker reducer
React re-invokes reducers during the render phase, so calling onChange
(and therefore a consumer's setState) inside datepickerReducer triggered
"Cannot update a component while rendering a different component"
warnings. onChange is now invoked from the actual Calendar/Input event
handlers, matching the pattern already used by DatePickerRange.

Fixes #1359
