+++ packages/datepickers/src/elements/DatePicker/utils/date-picker-reducer.ts
@@ -10,7 +10,6 @@ import { subMonths } from 'date-fns/subMonths';
 import { isValid } from 'date-fns/isValid';
 import { parse } from 'date-fns/parse';
 import { isBefore } from 'date-fns/isBefore';
-import { isSameDay } from 'date-fns/isSameDay';
 import { IDatePickerProps } from '../../../types';
 
 export interface IDatePickerState {
@@ -22,7 +21,7 @@ export interface IDatePickerState {
 /**
  * Parse string input value using current locale and date formats
  */
-function parseInputValue({
+export function parseInputValue({
   inputValue,
   customParseDate
 }: {
@@ -96,15 +95,11 @@ export const datepickerReducer =
   ({
     value,
     formatDate,
-    locale,
-    customParseDate,
-    onChange
+    locale
   }: {
     value?: Date;
     formatDate?: (date: Date) => string;
     locale: any;
-    customParseDate?: (inputValue: string) => Date;
-    onChange?: (date: Date) => void;
   }) =>
   (state: IDatePickerState, action: DatePickerAction): IDatePickerState => {
     switch (action.type) {
@@ -127,11 +122,6 @@ export const datepickerReducer =
       }
       case 'MANUALLY_UPDATE_INPUT': {
         const inputValue = action.value;
-        const currentDate = parseInputValue({ inputValue, customParseDate });
-
-        if (onChange && currentDate && isValid(currentDate) && !isSameDay(value!, currentDate)) {
-          onChange(currentDate);
-        }
 
         return { ...state, isOpen: true, inputValue };
       }
@@ -148,10 +138,6 @@ export const datepickerReducer =
       }
       case 'SELECT_DATE': {
         const inputValue = formatInputValue({ date: action.value, locale, formatDate });
-
-        if (onChange && action.value && isValid(action.value) && !isSameDay(value!, action.value)) {
-          onChange(action.value);
-        }
 
         return { ...state, isOpen: false, inputValue };
       }
